Additional changes per modargo's 2nd pass
And also broke down and just made a 'PMClaw' which is identical to Claw, except it uses the new Sharpen keyword.  Cause that speaks to the Ghost Claw mechanic and is otherwise a hidden thing on the base Claw card.  But it still looks like a Defect card!

Fixed an error with the code suggestion which removed the Array from ArrayList

ClawForOne extraneous array declaration removed

SearingClaw VFX plays multiple times based on upgrades

Fixed Knighthood missing an upgrade description line