Added "One Frame Mode" to the new config menu
Configurable toggle option to turn off all alternate card pack frames, for players who prefer a consistent look to the cards.

The config menu toggle added to all abstracts and other sources of card frame overrides throughout the project.  Toggle option properly saves and loads, but does require a restart to kick in.

Unfortunately required the Dimensional Cards to get legitimate card art and to render their images as normal, but so be it.