Strike a Bargain updated (and fixed a developer only hat menu crash)
Strike a Bargain reworked per discord chat - Reduces all attacks in hand to 0 instead of all attacks for the rest of the turn.

Added a more secure hat menu check for if an invalid hat has been chosen.  Currently crashes if you have a hat equipped from a pack that does not exist.