fix: Updating Unplayable Card Behavior (#15)
Unplayable cards now behave as follows:
Text half - always playable, any other effects on the card will work as if it was the normal card (ex. Normality limiting cards playable per turn to three)
Art half - Keeps the playability of the original card, but doesn't do anything else. So the art half of a Burn won't be playable, but also won't damage you at the end of the turn